# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: Laravel
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://brand_media.test'
paths:
...
    /api/v1/news:
        get:
            summary: 'List of news'
            description: ''
            parameters:
                -
                    in: query
                    name: page
                    description: 'Номер страницы с результатами выдачи'
                    example: 14
                    required: false
                    schema:
                        type: integer
                        description: 'Номер страницы с результатами выдачи'
                        example: 14
                -
                    in: query
                    name: sort
                    description: 'Поле для сортировки. По-умолчанию  ''id|desc'''
                    example: perferendis
                    required: false
                    schema:
                        type: string
                        description: 'Поле для сортировки. По-умолчанию  ''id|desc'''
                        example: perferendis
                -
                    in: query
                    name: search
                    description: 'Строка, которая должна содержаться в результатах выдачи'
                    example: blanditiis
                    required: false
                    schema:
                        type: string
                        description: 'Строка, которая должна содержаться в результатах выдачи'
                        example: blanditiis
                -
                    in: query
                    name: category_id
                    description: 'ID категории'
                    example: 13
                    required: false
                    schema:
                        type: integer
                        description: 'ID категории'
                        example: 13
                -
                    in: query
                    name: tag_id
                    description: 'ID тега'
                    example: 16
                    required: false
                    schema:
                        type: integer
                        description: 'ID тега'
                        example: 16
                -
                    in: query
                    name: author_id
                    description: 'ID автора'
                    example: 9
                    required: false
                    schema:
                        type: integer
                        description: 'ID автора'
                        example: 9
                -
                    in: query
                    name: from
                    description: 'Начало периода в формате ''YYYY-mm-dd HH:ii:ss'''
                    example: eius
                    required: false
                    schema:
                        type: string
                        description: 'Начало периода в формате ''YYYY-mm-dd HH:ii:ss'''
                        example: eius
                -
                    in: query
                    name: till
                    description: 'Окончание периода в формате ''YYYY-mm-dd HH:ii:ss'''
                    example: aut
                    required: false
                    schema:
                        type: string
                        description: 'Окончание периода в формате ''YYYY-mm-dd HH:ii:ss'''
                        example: aut
                -
                    in: query
                    name: status
                    description: 'Статус отображения (возможные значения visible, hidden)'
                    example: optio
                    required: false
                    schema:
                        type: string
                        description: 'Статус отображения (возможные значения visible, hidden)'
                        example: optio
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                401:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    message: Unauthenticated.
                                properties:
                                    message: { type: string, example: Unauthenticated. }
            tags:
                - News
            security: []
        post:
            summary: 'Create news'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - News
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'Название новости'
                                    example: qui
                                text:
                                    type: string
                                    description: 'optional Текст новости'
                                    example: ad
                                category_id:
                                    type: integer
                                    description: 'ID категории новости'
                                    example: 20
                                author_id:
                                    type: integer
                                    description: 'ID автора новости'
                                    example: 18
                                image:
                                    type: optional
                                    description: 'Картинка в base64'
                                    example: quia
                                'meta:title':
                                    type: string
                                    description: 'optional Содержимое тега <title>'
                                    example: totam
                                'meta:description':
                                    type: string
                                    description: 'optional Содержимое тега <meta name="description">'
                                    example: odit
                                tag_id:
                                    type: array
                                    description: 'Массив IDs тегов новости'
                                    example: quis
                                position:
                                    type: integer
                                    description: 'Позиция новости в плашке'
                                    example: 11
                                is_promo:
                                    type: integer
                                    description: 'Промо (возможные значения 1, 0)'
                                    example: 18
                            required:
                                - name
...